requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
autoflake==1.1
chardet==3.0.4
discord.py==0.16.12
flake8==3.5.0
mccabe==0.6.1
multidict==4.1.0
requests==2.18.4
pycodestyle==2.3.1
pyflakes==1.6.0
websockets==3.4
